Refactor command line options structure
