+++ Kragle/Kragle/Kragle.csproj
@@ -65,6 +65,13 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="CodeParser.cs" />
+    <Compile Include="ConsoleOptions\ParseSubOptions.cs" />
+    <Compile Include="ConsoleOptions\SubOptions.cs" />
+    <Compile Include="ConsoleOptions\OpenSubOptions.cs" />
+    <Compile Include="ConsoleOptions\ProjectsSubOptions.cs" />
+    <Compile Include="ConsoleOptions\UsersSubOptions.cs" />
+    <Compile Include="ConsoleOptions\Options.cs" />
+    <Compile Include="ConsoleOptions\ResetSubOptions.cs" />
     <Compile Include="CsvWriter.cs" />
     <Compile Include="FileStore.cs" />
     <Compile Include="Logger.cs" />
